+++ mfstools/rgetgoal.cc
@@ -1,0 +1,6 @@
+#include "getgoal.h"
+#include "rgetgoal.h"
+
+//const bool getGoalDeprecatedToolRegistered =
+  //      LizardTool::registerDeprecatedTool("mfsrgetgoal", new GetGoalTool("rgetgoal", true));
+
