+++ mfstools/rgetgoal.h
@@ -1,0 +1,7 @@
+#ifndef LIZARDFS_MFSTOOLS_RGETGOAL_H_
+#define LIZARDFS_MFSTOOLS_RGETGOAL_H_
+
+//static const bool getGoalDeprecatedToolRegistered;
+
+#endif
+
