+++ mfstools/rgettrashtime.cc
@@ -1,0 +1,5 @@
+#include "gettrashtime.h"
+#include "rgettrashtime.h"
+
+//static const bool getTrashTimeDeprecatedToolRegistered = LizardTool::registerDeprecatedTool("mfsrgettrashtime", new GetTrashTimeTool("rgettrashtime", true));
+
