+++ mfstools/rgettrashtime.h
@@ -1,0 +1,7 @@
+#ifndef LIZARDFS_MFSTOOLS_RGETTRASHTIME_H_
+#define LIZARDFS_MFSTOOLS_RGETTRASHTIME_H_
+
+//static const bool getTrashTimeDeprecatedToolRegistered;
+
+#endif
+
