+++ mfstools/rsetgoal.cc
@@ -1,0 +1,5 @@
+#include "setgoal.h"
+#include "rsetgoal.h"
+
+//const bool setGoalDeprecatedTool = LizardTool::registerDeprecatedTool( "mfsrsetgoal", new SetGoalTool("rsetgoal", true));
+
