+++ mfstools/rsetgoal.h
@@ -1,0 +1,7 @@
+#ifndef LIZARDFS_MFSTOOLS_RSETGOAL_H_
+#define LIZARDFS_MFSTOOLS_RSETGOAL_H_
+
+//static const bool setGoalDeprecatedTool;
+
+#endif //LIZARDFS_MFSTOOLS_RSETGOAL_H_
+
