+++ mfstools/rsettrashtime.cc
@@ -1,0 +1,5 @@
+#include "settrashtime.h"
+#include "rsettrashtime.h"
+
+//const bool setTrashTimeDeprecatedToolRegistered = LizardTool::registerDeprecatedTool("mfsrsettrashtime", new SetTrashTimeTool("rsettrashtime", true));
+
