+++ mfstools/rsettrashtime.h
@@ -1,0 +1,7 @@
+#ifndef LIZARDFS_MFSTOOLS_RSETTRASHTIME_H_
+#define LIZARDFS_MFSTOOLS_RSETTRASHTIME_H_
+
+//static const bool setTrashTimeDeprecatedToolRegistered;
+
+#endif //LIZARDFS_MFSTOOLS_RSETTRASHTIME_H_
+
