+++ mfstools/toolscommon.h
@@ -1,0 +1,29 @@
+#ifndef TOOLSCOMMON_H
+#define TOOLSCOMMON_H
+#include "lizardtools.h"
+#include "appendchunks.h"
+#include "checkfile.h"
+#include "deleattr.h"
+#include "makesnapshot.h"
+#include "setgoal.h"
+#include "getgoal.h"
+#include "rgetgoal.h"
+#include "settrashtime.h"
+#include "rsettrashtime.h"
+#include "rgettrashtime.h"
+#include "setquota.h"
+#include "seteattr.h"
+#include "gettrashtime.h"
+#include "getquota.h"
+#include "geteattr.h"
+#include "filerepair.h"
+#include "fileinfo.h"
+#include "dirinfo.h"
+#include "delquota.h"
+#include "lizardlink.h"
+
+
+
+static bool registerTools();
+
+#endif // TOOLSCOMMON_H
